eden_gradio/requirements.txt
torch
transformers
peft
Pillow
gradio
librosa
nltk